some more int->NSInteger and property stuff, some gc related fixes, send afterEvent to webif though it cannot be changed in the gui yet
git-svn-id: svn://svn.freaque.net/enigma2@426 a295cab0-163a-0410-857a-8207b47ee9d6

+++ Classes/ControlViewController.m
@@ -180,7 +180,7 @@
 
 - (void)volumeChanged:(id)sender
 {
-	[[RemoteConnectorObject sharedRemoteConnector] setVolume:(int)[(UISlider*)sender value]];
+	[[RemoteConnectorObject sharedRemoteConnector] setVolume:(NSInteger)[(UISlider*)sender value]];
 }
 
 - (BOOL)shouldAutorotateToInterfaceOrientation:(UIInterfaceOrientation)interfaceOrientation {
